Track clean re-usable in memory index pages in a LRU Map Add an auxiliary data structure to track LedgerEntryPages that are clean and not in use. This list is traversed to find available pages that can be reused when looking for a clean page. The auxiliary data structure is a LinkedHashMap so it naturally maintains the entries in access order providing LRU semantics.
There are no changes to the use of the nested concurrent maps - that data structure still continues to be the definitive map for looking up the LedgerEntryPages. There is no change to the logic that flushes the index.

I have removed the tracking of clean and dirty ledgers - I still retain some aspects of dirtyLedgers to cycle through the ledgers one by one as they are flushed to create clean pages.

RB_ID=97874