Hedwig multi-threaded readahead cache support:
This change introduces multiple-threads support for hedwig-server readahead cache,  each topic uses affinity of one thread in the threads pool. Further, guava caching is used to optimize for server work load, instead of the single-thread map in the past. (In single thread test cases, the performance improvement is not observable).

By default, (# of processors) threads are allocated to serve the readahead cache.

Note that it is possible (though unlikely) that stub entry can get evicted by LRU policy, the callback will be notified with a distinct exception (newly introduced).

The per-ledger entry tracking map was to allow entries consumed to be removed aggressively.
Given that guava cache already have the lru policy, which the new design leverage, it is less an issue.
Expire-after-access will adds overhead to the cache, and it does not seem worth to use for this purpose.

RB_ID=98825