BK: Fix flaky test on LedgerCloseTest
the test is to test ledger length should be consistent during close. so when ledger handle is being closed after erring out entries, the ledger handle length should be going to zero. but the ledger close is in background when erring out entries, so when the test tried to #openLedger, it might trigger a recover procedure, but there are no bookies available, so the test will be failed.

the fix is to start bookies after entries being erred out.

RB_ID=376555