BookKeeper Client: Delay Ensemble Change and ReadOnly Bookies When a bookie is marked readonly replace it from the ensemble proactively (even if delay ensemble change is enabled).
Although the bookie server always returned a separate exception and the client defined a separate exception, it was not translating the bookie error to the correct client side exception. That has been changed, however it should have no side effects as WriteException from BK client is not treated specially upstream.

RB_ID=549706