Checkpoint change (part 3): Change checkpoint logic from entrylog rolling trigger to periodic checkpointing
This change is majorly changing the way how bookie does checkpoint. (majorly porting fro OSS)

Originally, the checkpoint logic is driven by entry log file rotataion. we only issue checkpoint when a entry log file rotated.

This change is moving towards periodic checkpointing, which is driven by the SyncThread directly. And ledger storage will decide if it should checkpoint or not and return corresponding checkpoint that it finished.

- Move checkpoint source from SyncThread to Journal, as the journal is the actual source of checkpoints
- Make SyncThread use executor rather than Thread
- InterleaveLedgerStorage maintains a checkpoint and update each time entry log file is rotated.
- Checkpointing only happens when it is good to checkpoint (from performance perspective)

RB_ID=731976