Fix the bookkeeper script to start some servers.
The script had a broken test and was failing to see that the bookkeeper
jar was missing.

The test is fixed to be sure that we stop if we don't find a jar. Also,
we propose to the user to run `mvn package -DskipTests` if there's no
jar.

RB_ID=809974