A few updates to the start-stop script.
The script was using `ps -p` to test if a process is running. The `-p`
option is not available on all implementations of `ps` (e.g.:
busybox). Instead, use `kill -0`, which is more portable to verify if a
PID is running.

Reformat the code to use 2 spaces instead of a mix of 2 and 4 (2 was
dominant in the script, so I stick to that instead of 4).

Replace backticks with subshell, which is also more portable.

Fix a test in the script to start the server (checking if the jar file
exists).

RB_ID=816412